+++ apps/bot/src/domains/player/build-profil-view.ts
