+++ apps/bot/src/domains/player/interactions/profil-button.ts
@@ -2,8 +2,8 @@ import type { ButtonInteraction } from 'discord.js';
 
 import type { ButtonHandler } from '../../../discord/types.js';
 import { assertMenuOwner, parseIntegerArg, parseMenuOwnerDiscordId } from '../../../discord/utils/index.js';
+import { buildProfilView } from '../build-profil-view.js';
 import { PROFIL_BUTTON_NAME } from '../constants.js';
-import { buildProfilView } from '../profil-view.js';
 
 async function handle(interaction: ButtonInteraction, args: Array<string>): Promise<void> {
   const ownerDiscordId = parseMenuOwnerDiscordId(args[0]);
